added exception handling of aws service

+++ src/main/java/com/alkemy/ong/aws/AWSS3ServiceImpl.java
@@ -1,6 +1,8 @@
 package com.alkemy.ong.aws;
 
 import com.alkemy.ong.aws.utils.ValidImage;
+import com.amazonaws.AmazonServiceException;
+import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.PutObjectRequest;
@@ -11,6 +13,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
@@ -19,13 +22,13 @@ import java.io.IOException;
 import java.util.Date;
 import java.util.Locale;
 
+
 @RequiredArgsConstructor
 @Service
 public class AWSS3ServiceImpl implements IAWSS3Service {
 
     private final AmazonS3 s3client;
     private final MessageSource messageSource;
-    private static final Logger LOGGER = LoggerFactory.getLogger(AWSS3ServiceImpl.class);
 
     @Value("${amazonProperties.endpointUrl}")
     private String endpointUrl;
@@ -33,21 +36,27 @@ public class AWSS3ServiceImpl implements IAWSS3Service {
     private String bucketName;
 
     @Override
-    public String uploadFile(MultipartFile multipartFile) throws Exception{
-        String uploadErrorMessage = messageSource.getMessage("awss3.errorMessage.uploadFile", null, Locale.US);
+    public String uploadImage(MultipartFile multipartFile) throws IOException, SdkClientException{
+        String emptyFileErrorMessage = messageSource.getMessage("awss3.errorMessage.emptyFile", null, Locale.US);
+        String contentTypeErrorMessage = messageSource.getMessage("awss3.errorMessage.contentType", null, Locale.US);
         String fileUrl = "";
-        try {
-            File file = convertMultiPartToFile(multipartFile);
-            String fileName = generateFileName(multipartFile);
-            fileUrl = endpointUrl + "/" + bucketName + "/" + fileName;
-            uploadFileTos3bucket(fileName, file);
-            file.delete();
-        }catch (Exception e){
-            e.printStackTrace();
-            throw new Exception(uploadErrorMessage);
-        }
-        return fileUrl;
 
+        if(multipartFile.isEmpty()){
+            throw new IllegalArgumentException(emptyFileErrorMessage);
+        }
+
+        String contentType = multipartFile.getContentType();
+        if (!isSupportedContentType(contentType)) {
+            throw new IllegalArgumentException(contentTypeErrorMessage);
+        }
+
+        File file = convertMultiPartToFile(multipartFile);
+        String fileName = generateFileName(multipartFile);
+        fileUrl = endpointUrl + "/" + bucketName + "/" + fileName;
+        uploadFileTos3bucket(fileName, file);
+        file.delete();
+
+        return fileUrl;
     }
 
     private void uploadFileTos3bucket(String fileName, File file) {
@@ -65,6 +74,12 @@ public class AWSS3ServiceImpl implements IAWSS3Service {
         fos.write(file.getBytes());
         fos.close();
         return convFile;
+    }
+
+    private boolean isSupportedContentType(String contentType) {
+        return contentType.equals("image/png")
+                || contentType.equals("image/jpg")
+                || contentType.equals("image/jpeg");
     }
